XMLHttpRequest/abort-during-upload.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: abort() while sending data</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-abort()-method" data-tested-assertations="following-sibling::ol/li[4]/ol/li[7] following-sibling::ol/li[4]/ol/li[7]/ol/li[2] following-sibling::ol/li[4]/ol/li[7]/ol/li[3] following-sibling::ol/li[4]/ol/li[7]/ol/li[4]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      var result = [];
      var expected = ['progress on XHR', 'abort on XHR', 'loadend on XHR', 'progress on XHR Upload', 'abort on XHR Upload', 'loadend on XHR Upload']
      function logEvt (e) {
        var str = e.type+' on '
        str += e.target instanceof XMLHttpRequest ? 'XHR' : 'XHR Upload'
        result.push(str)
        console.log(str)
      }
      test.step(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "resources/delay.php?ms=1000")
        client.addEventListener('progress', logEvt)
        client.addEventListener('abort', logEvt)
        client.addEventListener('loadend', logEvt)
        client.upload.addEventListener('progress', logEvt)
        client.upload.addEventListener('abort', logEvt)
        client.upload.addEventListener('loadend', logEvt)
        client.upload.addEventListener('loadend', function () {
          test.step(function() {
            assert_equals(client.readyState, 0)
            assert_array_equals(result, expected)
            test.done()
          })
        })
        client.send((new Array(10000)).join('a'))
        client.abort();
      })
    </script>
  </body>
</html>
